ADK MCP/requirements.txt
# MCP Implementation Requirements
# Complete dependency list for the Model Context Protocol (MCP) implementation
# with Google ADK integration and FastAPI service

# Core MCP and Google ADK
google-adk==1.1.1
fastmcp==2.5.1
google-genai==0.8.1

# FastAPI Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic>=2.4.0
python-multipart>=0.0.6

# Security and Authentication
PyJWT==2.10.1
cryptography==45.0.5
requests==2.32.4

# Google Cloud Services
google-cloud-secret-manager==2.24.0
google-cloud-kms==3.5.1

# Environment and Configuration
python-dotenv==1.1.0

# Development and Utility
rich==13.9.4
jsonschema==4.23.0

# Production WSGI Server
gunicorn>=21.2.0

# Additional dependencies that may be needed
# asyncio-mqtt>=0.11.0  # For MQTT support
# websockets>=11.0.0    # For WebSocket support
# redis>=4.5.0          # For caching and session management
